+++ core/common-core/src/main/java/org/eclipse/edc/identityhub/DefaultServicesExtension.java
@@ -33,6 +33,7 @@ import org.eclipse.edc.identityhub.spi.keypair.store.KeyPairResourceStore;
 import org.eclipse.edc.identityhub.spi.participantcontext.store.ParticipantContextStore;
 import org.eclipse.edc.identityhub.spi.transformation.ScopeToCriterionTransformer;
 import org.eclipse.edc.identityhub.spi.verifiablecredentials.store.CredentialStore;
+import org.eclipse.edc.jsonld.spi.JsonLd;
 import org.eclipse.edc.jwt.signer.spi.JwsSignerProvider;
 import org.eclipse.edc.jwt.validation.jti.JtiValidationStore;
 import org.eclipse.edc.keys.spi.PrivateKeyResolver;
@@ -48,8 +49,16 @@ import org.eclipse.edc.spi.types.TypeManager;
 import org.eclipse.edc.token.spi.TokenValidationRulesRegistry;
 import org.eclipse.edc.verifiablecredentials.jwt.rules.JtiValidationRule;
 
+import java.net.URISyntaxException;
 import java.time.Clock;
 
+import static org.eclipse.edc.iam.identitytrust.spi.DcpConstants.DCP_CONTEXT_URL;
+import static org.eclipse.edc.iam.identitytrust.spi.DcpConstants.DSPACE_DCP_V_1_0_CONTEXT;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.VcConstants.DID_CONTEXT_URL;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.VcConstants.JWS_2020_URL;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.VcConstants.PRESENTATION_EXCHANGE_URL;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.VcConstants.PRESENTATION_SUBMISSION_URL;
+import static org.eclipse.edc.iam.verifiablecredentials.spi.VcConstants.W3C_CREDENTIALS_URL;
 import static org.eclipse.edc.identityhub.DefaultServicesExtension.NAME;
 import static org.eclipse.edc.identityhub.spi.verification.SelfIssuedTokenConstants.ACCESS_TOKEN_SCOPE_CLAIM;
 import static org.eclipse.edc.identityhub.spi.verification.SelfIssuedTokenConstants.DCP_PRESENTATION_ACCESS_TOKEN_CONTEXT;
@@ -58,7 +67,15 @@ import static org.eclipse.edc.identityhub.spi.verification.SelfIssuedTokenConsta
 
 @Extension(NAME)
 public class DefaultServicesExtension implements ServiceExtension {
+    public static final String PRESENTATION_EXCHANGE_V_1_JSON = "presentation-exchange.v1.json";
+    public static final String PRESENTATION_QUERY_V_08_JSON = "dcp.v08.json";
+    public static final String DSPACE_DCP_V_1_0_JSON_LD = "dcp.v1.0.jsonld";
+    public static final String PRESENTATION_SUBMISSION_V1_JSON = "presentation-submission.v1.json";
+    public static final String DID_JSON = "did.json";
+    public static final String JWS_2020_JSON = "jws2020.json";
+    public static final String CREDENTIALS_V_1_JSON = "credentials.v1.json";
 
+    
     public static final String NAME = "IdentityHub Default Services Extension";
     public static final long DEFAULT_REVOCATION_CACHE_VALIDITY_MILLIS = 15 * 60 * 1000L;
     static final String ACCESSTOKEN_JTI_VALIDATION_ACTIVATE = "edc.iam.accesstoken.jti.validation";
@@ -79,6 +96,8 @@ public class DefaultServicesExtension implements ServiceExtension {
     private Clock clock;
     @Inject
     private CriterionOperatorRegistry criterionOperatorRegistry;
+    @Inject
+    private JsonLd jsonLd;
 
     @Override
     public String name() {
@@ -99,6 +118,9 @@ public class DefaultServicesExtension implements ServiceExtension {
         } else {
             context.getMonitor().warning("JWT Token ID (\"jti\" claim) Validation is not active. Please consider setting '%s=true' for protection against replay attacks".formatted(ACCESSTOKEN_JTI_VALIDATION_ACTIVATE));
         }
+
+        // Setup API
+        cacheContextDocuments(getClass().getClassLoader());
     }
 
     @Provider(isDefault = true)
@@ -148,5 +170,18 @@ public class DefaultServicesExtension implements ServiceExtension {
         return new InMemoryHolderCredentialRequestStore(clock, criterionOperatorRegistry);
     }
 
+    private void cacheContextDocuments(ClassLoader classLoader) {
+        try {
+            jsonLd.registerCachedDocument(PRESENTATION_EXCHANGE_URL, classLoader.getResource(PRESENTATION_EXCHANGE_V_1_JSON).toURI());
+            jsonLd.registerCachedDocument(DCP_CONTEXT_URL, classLoader.getResource(PRESENTATION_QUERY_V_08_JSON).toURI());
+            jsonLd.registerCachedDocument(DSPACE_DCP_V_1_0_CONTEXT, classLoader.getResource(DSPACE_DCP_V_1_0_JSON_LD).toURI());
+            jsonLd.registerCachedDocument(DID_CONTEXT_URL, classLoader.getResource(DID_JSON).toURI());
+            jsonLd.registerCachedDocument(JWS_2020_URL, classLoader.getResource(JWS_2020_JSON).toURI());
+            jsonLd.registerCachedDocument(W3C_CREDENTIALS_URL, classLoader.getResource(CREDENTIALS_V_1_JSON).toURI());
+            jsonLd.registerCachedDocument(PRESENTATION_SUBMISSION_URL, classLoader.getResource(PRESENTATION_SUBMISSION_V1_JSON).toURI());
+        } catch (URISyntaxException e) {
+            throw new RuntimeException(e);
+        }
+    }
 
 }
