+++ core/common-core/src/main/resources/credentials.v1.json
