+++ core/common-core/src/main/resources/dcp.v08.json
