+++ core/common-core/src/main/resources/dcp.v1.0.jsonld
