+++ core/common-core/src/main/resources/did.json
