+++ core/common-core/src/main/resources/jws2020.json
