+++ core/common-core/src/main/resources/presentation-exchange.v1.json
