+++ core/common-core/src/main/resources/presentation-submission.v1.json
